Highly suggest new players set this up before playing.
For those of us that have been using an autoexec.cfg for years (or legacy from wc3 for example) + our own weird binds that make no sense but, are muscle memory so we don't change them it would be EXTREMELY HELPFUL
if you compiled all of the scripts, keybinds, and so forth into one "master file" so I can add the one's I need.  
When trying to copy paste scripts and binds from different locations causes a fail in-game.